feat: add nextui env command
