+++ src/constants/path.ts
@@ -1,1 +1,4 @@
+import { resolve } from 'path';
+
 export const ROOT = process.cwd();
+export const resolver = (path: string) => resolve(ROOT, path);
