+++ src/helpers/logger.ts
@@ -1,5 +1,7 @@
 /* eslint-disable no-console */
 
+import type { SAFE_ANY } from './type';
+
 import chalk from 'chalk';
 import { default as _gradientString } from 'gradient-string';
 
@@ -7,6 +9,9 @@ export const defaultColors = ['#FF1CF7', '#B249F8'] as const;
 
 export const gradientString = _gradientString;
 
+const logPrefix = gradientString(...defaultColors)('NextUI CLI:');
+
+type PrefixLogType = Extract<keyof typeof Logger, 'error' | 'gradient' | 'info' | 'log' | 'warn'>;
 export class Logger {
   constructor() {}
 
@@ -31,6 +36,10 @@ export class Logger {
     options?: { colors?: tinycolor.ColorInput[] }
   ) {
     this.log(gradientString(...(options?.colors ?? defaultColors))(String(content)));
+  }
+
+  static prefix(type: PrefixLogType, ...args: SAFE_ANY) {
+    return this[type](logPrefix, ...args);
   }
 
   static newLine(lines?: number) {
